+++ supriya/ugens/reverb.py
@@ -18,12 +18,7 @@ class FreeVerb(UGen):
 
     """
 
-    ### CLASS VARIABLES ###
-
-    __documentation_section__ = "Reverb UGens"
-
     _ordered_input_names = collections.OrderedDict(
         [("source", None), ("mix", 0.33), ("room_size", 0.5), ("damping", 0.5)]
     )
-
     _valid_calculation_rates = (CalculationRate.AUDIO,)
